+++ packages/grafana-ui/src/unstable.ts
@@ -8,3 +8,5 @@
  * Once mature, they will be moved to the main export, be available to plugins, and
  * be subject to the standard policies
  */
+
+export * from './utils/skeleton';
