+++ packages/grafana-ui/src/utils/index.ts
@@ -7,7 +7,6 @@ export * from './slate';
 export * from './dataLinks';
 export * from './tags';
 export * from './scrollbar';
-export * from './skeleton';
 export * from './table';
 export * from './measureText';
 export * from './useForceUpdate';
