+++ public/app/features/admin/AdminOrgsTable.tsx
@@ -3,7 +3,8 @@ import React, { useState } from 'react';
 import Skeleton from 'react-loading-skeleton';
 
 import { GrafanaTheme2 } from '@grafana/data';
-import { Button, ConfirmModal, SkeletonComponent, useStyles2, attachSkeleton } from '@grafana/ui';
+import { Button, ConfirmModal, useStyles2 } from '@grafana/ui';
+import { SkeletonComponent, attachSkeleton } from '@grafana/ui/src/unstable';
 import { contextSrv } from 'app/core/core';
 import { AccessControlAction, Organization } from 'app/types';
