+++ public/app/features/library-panels/components/LibraryPanelCard/LibraryPanelCard.tsx
@@ -4,7 +4,8 @@ import Skeleton from 'react-loading-skeleton';
 
 import { GrafanaTheme2 } from '@grafana/data';
 import { config } from '@grafana/runtime';
-import { Icon, Link, SkeletonComponent, useStyles2, attachSkeleton } from '@grafana/ui';
+import { Icon, Link, useStyles2 } from '@grafana/ui';
+import { SkeletonComponent, attachSkeleton } from '@grafana/ui/src/unstable';
 import { getPanelPluginNotFound } from 'app/features/panel/components/PanelPluginError';
 import { PanelTypeCard } from 'app/features/panel/components/VizTypePicker/PanelTypeCard';
