+++ public/app/features/panel/components/VizTypePicker/PanelTypeCard.tsx
@@ -4,7 +4,8 @@ import Skeleton from 'react-loading-skeleton';
 
 import { GrafanaTheme2, isUnsignedPluginSignature, PanelPluginMeta, PluginState } from '@grafana/data';
 import { selectors } from '@grafana/e2e-selectors';
-import { IconButton, PluginSignatureBadge, SkeletonComponent, useStyles2, attachSkeleton } from '@grafana/ui';
+import { IconButton, PluginSignatureBadge, useStyles2 } from '@grafana/ui';
+import { SkeletonComponent, attachSkeleton } from '@grafana/ui/src/unstable';
 import { PluginStateInfo } from 'app/features/plugins/components/PluginStateInfo';
 
 interface Props {
