+++ public/app/features/playlist/PlaylistCard.tsx
@@ -3,16 +3,8 @@ import React from 'react';
 import Skeleton from 'react-loading-skeleton';
 
 import { GrafanaTheme2 } from '@grafana/data';
-import {
-  Button,
-  Card,
-  LinkButton,
-  ModalsController,
-  SkeletonComponent,
-  Stack,
-  useStyles2,
-  attachSkeleton,
-} from '@grafana/ui';
+import { Button, Card, LinkButton, ModalsController, Stack, useStyles2 } from '@grafana/ui';
+import { attachSkeleton, SkeletonComponent } from '@grafana/ui/src/unstable';
 import { t, Trans } from 'app/core/internationalization';
 import { contextSrv } from 'app/core/services/context_srv';
 import { DashNavButton } from 'app/features/dashboard/components/DashNav/DashNavButton';
