+++ public/app/features/playlist/PlaylistPageList.tsx
@@ -2,7 +2,8 @@ import { css } from '@emotion/css';
 import React from 'react';
 
 import { GrafanaTheme2 } from '@grafana/data';
-import { SkeletonComponent, useStyles2, attachSkeleton } from '@grafana/ui';
+import { useStyles2 } from '@grafana/ui';
+import { SkeletonComponent, attachSkeleton } from '@grafana/ui/src/unstable';
 
 import { PlaylistCard } from './PlaylistCard';
 import { Playlist } from './types';
