+++ public/app/features/serviceaccounts/components/ServiceAccountsListItem.tsx
@@ -3,7 +3,8 @@ import React, { memo } from 'react';
 import Skeleton from 'react-loading-skeleton';
 
 import { GrafanaTheme2, OrgRole } from '@grafana/data';
-import { Button, Icon, IconButton, type SkeletonComponent, Stack, useStyles2, attachSkeleton } from '@grafana/ui';
+import { Button, Icon, IconButton, Stack, useStyles2 } from '@grafana/ui';
+import { SkeletonComponent, attachSkeleton } from '@grafana/ui/src/unstable';
 import { UserRolePicker } from 'app/core/components/RolePicker/UserRolePicker';
 import { contextSrv } from 'app/core/core';
 import { OrgRolePicker } from 'app/features/admin/OrgRolePicker';
